feat(readme-generator): move the readme file outside of the script and ignore deprecated apps
